Merge csdb/README.md with README.md and revise
The resulting README.md in the repo root is a combination of the
original README.md and csdb/README.md, with most of the content coming
from the latter. csdb/README.md was then removed.

This was done as the first step of making the csdb collection the *only*
collection in the repo by moving all its content out of the csdb
directory.