Remove conditional compilation from array tests
Previously, the code that used generics in TestUArraysCatSnippets was
not hidden from compilers that don't support generics via conditional
compilation. Since the tests are only designed for use with Delphi XE
and later, there is no need for this, so all the conditional compilation
directives were removed from the test unit.