Add overflow checking to PowNZN function
Updated PowNZN function in Mathematics category to check for integer
overflow and to raise an EOverflow exception when detected.

This change was required to trap errors when implementing DigitPowerSum, which
will call PowNZN.

Updated maths.ini re requirement for SysUtils unit in PowNZN and to updated
description to reference the newly added exception.